requirements.txt
discord.py
ccxt
pandas
python-dotenv
aiohttp
sqlalchemy
aiosqlite
scipy
numpy
mplfinance
matplotlib
